requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
djangorestframework==3.13.1
idna==3.3
pycodestyle==2.8.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
